+++ tarantool/src/cbus/sync/std.rs
@@ -310,7 +310,7 @@ mod tests {
 
     #[crate::test(tarantool = "crate")]
     pub fn single_producer() {
-        let mut cbus_fiber = run_cbus_endpoint("std_single_producer");
+        let cbus_fiber_id = run_cbus_endpoint("std_single_producer");
 
         let cap = NonZeroUsize::new(10).unwrap();
         let (tx, rx) = sync::std::channel("std_single_producer", cap);
@@ -335,12 +335,12 @@ mod tests {
             YieldResult::Yielded((0..1000).collect::<Vec<_>>())
         );
         thread.join().unwrap();
-        cbus_fiber.cancel();
+        assert!(fiber::cancel(cbus_fiber_id));
     }
 
     #[crate::test(tarantool = "crate")]
     pub fn single_producer_lock() {
-        let mut cbus_fiber = run_cbus_endpoint("std_single_producer_lock");
+        let cbus_fiber_id = run_cbus_endpoint("std_single_producer_lock");
 
         static SEND_COUNTER: AtomicU64 = AtomicU64::new(0);
 
@@ -368,7 +368,7 @@ mod tests {
         assert_eq!((0..100).collect::<Vec<_>>(), recv_results);
 
         thread.join().unwrap();
-        cbus_fiber.cancel();
+        assert!(fiber::cancel(cbus_fiber_id));
     }
 
     #[crate::test(tarantool = "crate")]
@@ -377,7 +377,7 @@ mod tests {
         // receiver part. Previously, when the receiver was drop after sender, [`Fiber::Cond`] release outside the tx thread
         // and segfault is occurred.
 
-        let mut cbus_fiber = run_cbus_endpoint("std_drop_rx_before_tx");
+        let cbus_fiber_id = run_cbus_endpoint("std_drop_rx_before_tx");
         let cap = NonZeroUsize::new(1000).unwrap();
         let (tx, rx) = sync::std::channel("std_drop_rx_before_tx", cap);
 
@@ -394,12 +394,12 @@ mod tests {
         drop(rx);
         thread.join().unwrap();
 
-        cbus_fiber.cancel();
+        assert!(fiber::cancel(cbus_fiber_id));
     }
 
     #[crate::test(tarantool = "crate")]
     pub fn tx_disconnect() {
-        let mut cbus_fiber = run_cbus_endpoint("std_tx_disconnect");
+        let cbus_fiber_id = run_cbus_endpoint("std_tx_disconnect");
 
         let cap = NonZeroUsize::new(1).unwrap();
         let (tx, rx) = sync::std::channel("std_tx_disconnect", cap);
@@ -414,12 +414,12 @@ mod tests {
         assert!(matches!(rx.receive(), Err(RecvError::Disconnected)));
 
         thread.join().unwrap();
-        cbus_fiber.cancel();
+        assert!(fiber::cancel(cbus_fiber_id));
     }
 
     #[crate::test(tarantool = "crate")]
     pub fn rx_disconnect() {
-        let mut cbus_fiber = run_cbus_endpoint("std_rx_disconnect");
+        let cbus_fiber_id = run_cbus_endpoint("std_rx_disconnect");
 
         let cap = NonZeroUsize::new(1).unwrap();
         let (tx, rx) = sync::std::channel("std_rx_disconnect", cap);
@@ -435,13 +435,13 @@ mod tests {
         drop(rx);
 
         thread.join().unwrap();
-        cbus_fiber.cancel();
+        assert!(fiber::cancel(cbus_fiber_id));
     }
 
     #[crate::test(tarantool = "crate")]
     pub fn multiple_producer() {
         const MESSAGES_PER_PRODUCER: i32 = 10_000;
-        let mut cbus_fiber = run_cbus_endpoint("std_multiple_producer");
+        let cbus_fiber_id = run_cbus_endpoint("std_multiple_producer");
 
         let cap = NonZeroUsize::new(10).unwrap();
         let (tx, rx) = sync::std::channel("std_multiple_producer", cap);
@@ -466,13 +466,13 @@ mod tests {
         jh1.join().unwrap();
         jh2.join().unwrap();
         jh3.join().unwrap();
-        cbus_fiber.cancel();
+        assert!(fiber::cancel(cbus_fiber_id));
     }
 
     #[crate::test(tarantool = "crate")]
     pub fn multiple_producer_lock() {
         const MESSAGES_PER_PRODUCER: i32 = 100;
-        let mut cbus_fiber = run_cbus_endpoint("std_multiple_producer_lock");
+        let cbus_fiber_id = run_cbus_endpoint("std_multiple_producer_lock");
 
         let cap = NonZeroUsize::new(10).unwrap();
         let (tx, rx) = sync::std::channel("std_multiple_producer_lock", cap);
@@ -506,6 +506,6 @@ mod tests {
         jh1.join().unwrap();
         jh2.join().unwrap();
         jh3.join().unwrap();
-        cbus_fiber.cancel();
+        assert!(fiber::cancel(cbus_fiber_id));
     }
 }