+++ tests/src/latch.rs
@@ -1,34 +1,28 @@
 use std::rc::Rc;
 use std::time::Duration;
 
-use tarantool::fiber::{sleep, Fiber, Latch};
+use tarantool::fiber;
 
 pub fn latch_lock() {
-    let latch = Rc::new(Latch::new());
+    let latch = Rc::new(fiber::Latch::new());
 
-    let mut fiber = Fiber::new("test_fiber", &mut |_| {
+    let fiber = fiber::start(|| {
         let _lock = latch.lock();
-        sleep(Duration::from_millis(10));
-        0
+        fiber::sleep(Duration::from_millis(10));
     });
 
-    fiber.set_joinable(true);
-    fiber.start(());
     latch.lock();
     fiber.join();
 }
 
 pub fn latch_try_lock() {
-    let latch = Rc::new(Latch::new());
+    let latch = Rc::new(fiber::Latch::new());
 
-    let mut fiber = Fiber::new("test_fiber", &mut |_| {
+    let fiber = fiber::start(|| {
         let _lock = latch.lock();
-        sleep(Duration::from_millis(10));
-        0
+        fiber::sleep(Duration::from_millis(10));
     });
 
-    fiber.set_joinable(true);
-    fiber.start(());
     assert!(latch.try_lock().is_none());
 
     fiber.join();